koreabaseball-backend/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
sqlmodel==0.0.22
httpx==0.27.2
Scrapy==2.11.2
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.9.0
argon2-cffi==23.1.0